backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
google-generativeai
supabase
psycopg2-binary
python-dotenv
selenium
